requirements.txt
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
drf-extensions==0.6.0
drf-flex-fields==0.8.0
pendulum==2.1.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
pytzdata==2019.3
six==1.14.0
sqlparse==0.3.1
